Fix fetcher cache reduction
When artifact size is smaller than expected, we want to reduce recorded
cache usage.

To do it, we actually compute the delta as an off_t (signed). If delta is
negative, we reclaim the extra space as it's not really used.

This attempt was failing because computed delta is negative,
and passed as Bytes to releaseSpace, which is uint32_t
(unsigned). In most cases, casting a negative value into an uint will
just give an unrelated value due to sign bit disappearing.

By passing the positive value to Bytes(), we have a safer cast.

JIRA: MESOS-5292